feature#36 Tabrow 리플 제거 & 컴포넌트 분리
